requirements.txt
asgiref==3.5.2
backports.weakref==1.0.post1
Django==4.1.2
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.2.0
pycparser==2.21
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
whitenoise==6.2.0